requirements.txt
albumentations
ipykernel
matplotlib
jupyter
torch
torchvision
pandas
pyyaml
scikit-learn